# walredo: make request_redo() an async fn